test: isolate timerService and scheduleNextRound
